src/models/qlib_/requirements.txt
pandas
numpy
scikit-learn
mlflow
pyqlib
xgboost
catboost
lightgbm
